requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==4.0.3
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
build==1.4.4
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
chromadb==1.5.8
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
curl_cffi==0.13.0
dataclasses-json==0.6.7
defusedxml==0.7.1
duckduckgo_search==8.1.1
durationpy==0.10
exceptiongroup==1.3.1
fastapi==0.128.8
filelock==3.19.1
flatbuffers==25.12.19
fonttools==4.60.2
fpdf2==2.8.4
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.10.0
googleapis-common-protos==1.74.0
greenlet==3.2.4
grpcio==1.80.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.8.0
humanfriendly==10.0
idna==3.13
importlib_metadata==8.7.1
importlib_resources==6.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==0.3.28
langchain-chroma==0.2.6
langchain-community==0.3.31
langchain-core==0.3.84
langchain-ollama==0.3.10
langchain-text-splitters==0.3.11
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
lxml==6.1.0
markdown-it-py==3.0.0
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
multitasking==0.0.12
mypy_extensions==1.1.0
numpy==2.0.2
oauthlib==3.3.1
ollama==0.6.1
onnxruntime==1.19.2
opentelemetry-api==1.41.0
opentelemetry-exporter-otlp-proto-common==1.41.0
opentelemetry-exporter-otlp-proto-grpc==1.41.0
opentelemetry-proto==1.41.0
opentelemetry-sdk==1.41.0
opentelemetry-semantic-conventions==0.62b0
orjson==3.11.5
ormsgpack==1.11.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
peewee==4.0.4
pillow==11.3.0
platformdirs==4.4.0
primp==0.15.0
propcache==0.4.1
protobuf==6.33.6
pybase64==1.4.3
pycparser==2.23
pydantic==2.13.3
pydantic-settings==2.11.0
pydantic_core==2.46.3
Pygments==2.20.0
pypdf==6.10.2
PyPika==0.51.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.27.1
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
starlette==0.49.3
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.2
tomli==2.4.1
tqdm==4.67.3
typer==0.23.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.39.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
yfinance==1.2.0
zipp==3.23.1
zstandard==0.25.0
